create S3NonBlockingRouterTest for S3
